# definitions/openapi.yaml
openapi: 3.0.0
info:
  title: Event Booking API
  description: API for managing users, events, and bookings.
  version: 1.0.0
paths:
...
  /bookings:
    post:
      tags: 
        - Bookings
      summary: Create a booking
      operationId: createBooking
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: 
        - Bookings
      summary: Retrieve all bookings
      operationId: getBookings
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token for authentication
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrichedBookingResponse'
components:
  schemas:
...
    Event:
      type: object
      required:
        - id
        - title
        - startDateTime
        - endDateTime
        - hostId
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        hostId:
          type: string
    CreateBookingRequest:
      type: object
      required:
        - eventId
      properties:
        eventId:
          type: string
    Booking:
      type: object
      required:
        - id
        - userId
        - eventId
      properties:
        id:
          type: string
        userId:
          type: string
        eventId:
          type: string
    EnrichedBookingResponse:
      type: object
      required:
        - id
        - userId
        - eventId
        - event
      properties:
        id:
          type: string
        userId:
          type: string
        eventId:
          type: string
        event:
          $ref: '#/components/schemas/Event'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: 'Invalid credentials'