# src/models/bookings.ts
import { DataTypes, Model } from 'sequelize';
import { sequelize } from '../database';

export class Booking extends Model {
  declare id: string;
  declare userId: string;
  declare eventId: string;
}

Booking.init(
  {
    id: {
      type: DataTypes.UUID,
      defaultValue: DataTypes.UUIDV4,
      primaryKey: true,
    },
    userId: {
      type: DataTypes.UUID,
      allowNull: false,
    },
    eventId: {
      type: DataTypes.UUID,
      allowNull: false,
    },
  },
  {
    sequelize,
    modelName: 'Booking',
    tableName: 'bookings',
    timestamps: true,
  }
);

export default Booking;